+++ packages/core/src/js/integrations/graphql.ts
@@ -2,7 +2,7 @@ import type { Integration } from '@sentry/core';
 
 import { graphqlClientIntegration as browserGraphqlClientIntegration } from '@sentry/browser';
 
-interface GraphQLReactNativeIntegrationOptions {
+export interface GraphQLReactNativeIntegrationOptions {
   endpoints: Array<string | RegExp>;
 }